backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.1
google-generativeai==0.7.0
python-dotenv==1.0.1
